removing section 3
using updaters is different from enabling a script. Example covered in
this section does not correspond to how an updater works and actually,
example would not work properly either if a new script was created
instead of using a script test_script.sh that was whitelisted in the
previous section.